Refactor pytest fixtures, add tests for hubstorage client
