+++ tests/hubstorage/conftest.py
@@ -12,6 +12,8 @@ from requests import HTTPError
 from scrapinghub import HubstorageClient
 from scrapinghub.hubstorage.utils import urlpathjoin
 
+from ..conftest import request_accept_header_matcher
+
 
 TEST_PROJECT_ID = "2222222"
 TEST_SPIDER_NAME = 'hs-test-spider'
@@ -44,7 +46,10 @@ class VCRGzipSerializer(object):
 
 my_vcr = vcr.VCR(cassette_library_dir=VCR_CASSETES_DIR, record_mode='once')
 my_vcr.register_serializer('gz', VCRGzipSerializer())
+my_vcr.register_matcher('accept_header', request_accept_header_matcher)
 my_vcr.serializer = 'gz'
+my_vcr.match_on = ('method', 'scheme', 'host', 'port',
+                   'path', 'query', 'accept_header')
 
 
 def pytest_configure(config):
@@ -105,9 +110,14 @@ def setup_vcrpy(request, hsproject):
     # generates names like "test_module/test_function.yaml"
     # otherwise it uses current function name (setup_vcrpy) for all tests
     # other option is to add vcr decorator to each test separately
-    cassette_name = '{}/{}.gz'.format(
+    serializer_suffix = ''
+    if ('json_and_msgpack' in request.fixturenames and
+            request.getfixturevalue('json_and_msgpack') == 'json'):
+        serializer_suffix = '-json'
+    cassette_name = '{}/{}{}.gz'.format(
         request.function.__module__.split('.')[-1],
-        request.function.__name__
+        request.function.__name__,
+        serializer_suffix
     )
     if is_using_real_services(request):
         remove_all_jobs(hsproject)
